Update bclose function to work well with Fugitive diffs.
My normal workflow is:
 - <leader>tt - Open nerdtree
 - Open file
 - <leader>gs - Open git status
 - Highlight file in status
 - dp - Diff patch of file
 - <leader>bd - Close diffpatch.

The issue was the <leader>bd (which calss Bclose) would give an annoying
warning saying the buffer wasn't actually deleted because there was
never a buffer.  A better way of doing this would be `:Gw`, which
applies a Fugitive git diffpatch (not the best thing, but the best I could find).
So if filetype is git (which it is for fugitive git diffpatches) Run :Gw
instead of the normal buffer delete commands.